Fix Android build

+++ shlr/tcc/tcc.h
@@ -237,7 +237,7 @@ typedef struct SValue {
 typedef struct Sym {
     int v;    /* symbol token */
     char *asm_label;    /* associated asm label */
-    long r;    /* associated register */
+    size_t r;    /* associated register */
     union {
         long long c;    /* associated number */
         int *d;   /* define token stream */
